Add suffix core to core namespace
