+++ core/src/main/java/dev/jdtech/jellyfin/dialogs/ErrorDialogFragment.kt
@@ -5,7 +5,7 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.fragment.app.DialogFragment
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import dev.jdtech.jellyfin.R
+import dev.jdtech.jellyfin.core.R
 import dev.jdtech.jellyfin.utils.serializable
 import java.io.Serializable
 import java.lang.IllegalStateException
